+++ src/main/java/fr/estecka/variantscit/commands/ModuleCommands.java
@@ -6,7 +6,6 @@ import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.mojang.brigadier.suggestion.Suggestions;
 import com.mojang.brigadier.suggestion.SuggestionsBuilder;
 import fr.estecka.variantscit.VariantsCitMod;
-import fr.estecka.variantscit.itemdata.transforms.impl.LogTransform;
 import fr.estecka.variantscit.modules.IBakedModule;
 import fr.estecka.variantscit.modules.cache.ICacheKey;
 import fr.estecka.variantscit.reload.EModuleHook;
@@ -103,15 +102,17 @@ extends CommandUtil
 	}
 
 	static private int Execute(CommandContext<FabricClientCommandSource> context, IModuleCommand command) throws CommandSyntaxException {
-		EModuleHook moduleContext = getModuleHook(context, HOOK_ARG);
+		EModuleHook hook = getModuleHook(context, HOOK_ARG);
 		ResourceLocation moduleId = context.getArgument(MODULE_ARG, ResourceLocation.class);
 		MetaModule meta = VariantsCitMod.GetModules().GetMeta(moduleId);
-		IBakedModule module = meta.bakedModules().get(moduleContext);
+		if (meta == null)
+			return Error(context, "No such module: "+moduleId);
 
+		IBakedModule module = meta.bakedModules().get(hook);
 		if (module == null)
-			return Error(context, "No such module: "+moduleContext+" "+moduleId);
+			return Error(context, "No hook "+hook+" for module "+moduleId);
 
-		CommandLogger logger = new CommandLogger(context, moduleContext, meta, meta.modelPrefix());
+		CommandLogger logger = new CommandLogger(context, hook, meta, meta.modelPrefix());
 		return command.Execute(context, logger, module);
 	}